HW04_prepared

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserMealRepositoryImpl.java
@@ -36,65 +36,57 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
     @Autowired
     private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
-    private SimpleJdbcInsert insertMeal;
-
-    public void setJdbcTemplate(JdbcTemplate jdbcTemplate){ this.jdbcTemplate = jdbcTemplate;}
+    private SimpleJdbcInsert insertUserMeal;
 
     @Autowired
     public JdbcUserMealRepositoryImpl(DataSource dataSource) {
-        this.insertMeal = new SimpleJdbcInsert(dataSource)
-                .withTableName("MEALS")
-                .usingGeneratedKeyColumns("meal_id");
+        this.insertUserMeal = new SimpleJdbcInsert(dataSource)
+                .withTableName("meals")
+                .usingGeneratedKeyColumns("id");
     }
 
     @Override
     public UserMeal save(UserMeal userMeal, int userId) {
         MapSqlParameterSource map = new MapSqlParameterSource()
-                .addValue("meal_id", userMeal.getId())
+                .addValue("id", userMeal.getId())
                 .addValue("description", userMeal.getDescription())
-                .addValue("calories",userMeal.getCalories())
-                .addValue("dateTime", userMeal.getDateTime());
-        if(userMeal.isNew()){
-            Number newKey = insertMeal.executeAndReturnKey(map);
-            userMeal.setId(newKey.intValue());
+                .addValue("calories", userMeal.getCalories())
+                .addValue("date_time", userMeal.getDateTime())
+                .addValue("user_id", userId);
+
+        if (userMeal.isNew()) {
+            Number newId = insertUserMeal.executeAndReturnKey(map);
+            userMeal.setId(newId.intValue());
         } else {
-            namedParameterJdbcTemplate.update(
-                    "UPDATE meals SET description=:description, calories=:calories, datetime=:dateTime," +
-                            "WHERE meal_id=:meal_id", map);
+            if (namedParameterJdbcTemplate.update(
+                    "UPDATE meals SET description=:description, calories=:calories, date_time=:date_time " +
+                            " WHERE id=:id AND user_id=:user_id", map) == 0) {
+                return null;
+            }
         }
         return userMeal;
     }
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id) != 0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=?", id, userId) != 0;
     }
 
-    @Override
-    public boolean delete(int userId){
-        return jdbcTemplate.update("DELETE FROM meals WHERE id=?", userId) !=0;
-    }
     @Override
     public UserMeal get(int id, int userId) {
-        List<UserMeal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=?", ROW_MAPPER, id);
-        return DataAccessUtils.singleResult(meals);
+        List<UserMeal> userMeals = jdbcTemplate.query(
+                "SELECT * FROM meals WHERE id = ? AND user_id = ?", ROW_MAPPER, id, userId);
+        return DataAccessUtils.singleResult(userMeals);
     }
-
-    @Override
     public List<UserMeal> getAll(int userId) {
-        return jdbcTemplate.query("SELECT * FROM meals ORDER BY dateTime ", ROW_MAPPER);
+        return jdbcTemplate.query(
+                "SELECT * FROM meals WHERE user_id=? ORDER BY date_time DESC", ROW_MAPPER, userId);
     }
 
     @Override
     public List<UserMeal> getBetween(LocalDateTime startDate, LocalDateTime endDate, int userId) {
-        Map<String, Object> map = new HashMap<>();
-        map.put("startDate", startDate);
-        map.put("endDate", endDate);
-        map.put("id", userId);
-        return namedParameterJdbcTemplate.queryForList(
-                "UPDATE meals SET description=:description, calories=:calories, datetime=:dateTime," +
-                        "WHERE meal_id=:meal_id", map, UserMeal.class);
-
-
+        return jdbcTemplate.query(
+                "SELECT * FROM meals WHERE user_id=?  AND date_time BETWEEN  ? AND ? ORDER BY date_time DESC",
+                ROW_MAPPER, userId, startDate, endDate);
     }
 }
